Started work on the maze solver
Solve button now triggers a unique function in maze.js.

The maze is now a class for easier use.

The maze is stored as a tree

maze width and height are now linked together; renamed to size.

player movement code is not reduntant anymore

The maze can now be solved using the tree representation of said maze.