Alot has changed. I updated the CSS and HTML for the site and added a new project (quadTree visualizer). I will go into it below
css/index.css -
Changed the site font, and layout of projects. Added ability to scroll between projects and centered the projects correctly

css/prism.css and js/prism.js -
these are for displaying the relavant code for each project besides each project. They look neat right now, but I might just remove them later if I dont like the code being visible, or if I find a better way to display the code.

index.html -
renamed some of the elements in junction with css/index.css and added formating for the new project (quadtree) + a test for the code of the quad tree next to it. I hate the plaintext code, however it seems to be the best way so far, but it makes the index.html file ugly.

js/maze.js -
with newer knowledge of how JS works, I changed almost all of the 'var' declarations to 'let' declarations. I updated some of the comments as well. Some variable names have changed to be more descriptive.

js/QuadTree.js -
New project! This is an interactive quad tree visualizer. A mouse click on the quadTree canvas will create a point  and push it to the quad tree. The quad tree is a spacial partitioning structure. The quad tree efficiently stores points, this particular quad tree can store a maximum of one point per leaf node.