added about me page with hyperlinks to github, linkedin, and this portfolio. Added new functionality to the quad tree visualizer, and added/edited the code for both the maze and the quad tree.
index.html : This is now the homepage html file instead of the portfolio html file.

js/index.js : moved classes used by both projects into a common file. This includes the Rectangle class, node class, and point class.

js/maze.js : redid the recursion algorithm and removed the maze class. The maze is now represented as a tree, rather than a 2D array of nodes. This allows control over the size of the maze. The new recursion algorithm doesn't make a maze that can be represented as a binary tree, instead it directly creates a tree which is not limited to two children per node. The maze display functions were also changed to represent walls as thin lines rather than filled cells. The maze solving algorithm was also fixed so that it no longer occasionally doesn't complete the maze. The solved maze path visualizer was changed to be a solid line.

js/QuadTree.js : added functionality to the quad tree to support a query within a rectangle, returning all points within the range. Removed individual point placement in favor of a random Gaussian distribution of 500 points on generation. This is to display the functionality of the rectangular query function. Dragging the mouse over the quad tree's canvas highlights points within a rectangle.

portfolio.html : this is the original html file, now utilized for only the projects.